+++ _test_null_safety/test/common/message.dart
@@ -1,0 +1,6 @@
+// Copyright (c) 2020, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+// Uses `?` to ensure null safety is enabled
+final String? message = 'hello';