+++ build_web_compilers/lib/src/dart2js_bootstrap.dart
@@ -10,8 +10,10 @@ import 'package:archive/archive.dart';
 import 'package:build/build.dart';
 import 'package:build/experiments.dart';
 import 'package:build_modules/build_modules.dart';
+import 'package:build_modules/src/workers.dart';
 import 'package:glob/glob.dart';
 import 'package:path/path.dart' as p;
+import 'package:pool/pool.dart';
 import 'package:scratch_space/scratch_space.dart';
 
 import 'platforms.dart';
@@ -22,6 +24,11 @@ import 'web_entrypoint_builder.dart';
 /// If [skipPlatformCheck] is `true` then all `dart:` imports will be
 /// allowed in all packages.
 Future<void> bootstrapDart2Js(BuildStep buildStep, List<String> dart2JsArgs,
+        {bool skipPlatformCheck}) =>
+    _resourcePool.withResource(() => _bootstrapDart2Js(buildStep, dart2JsArgs,
+        skipPlatformCheck: skipPlatformCheck));
+
+Future<void> _bootstrapDart2Js(BuildStep buildStep, List<String> dart2JsArgs,
     {bool skipPlatformCheck}) async {
   skipPlatformCheck ??= false;
   var dartEntrypointId = buildStep.inputId;
@@ -66,20 +73,29 @@ https://github.com/dart-lang/build/blob/master/docs/faq.md#how-can-i-resolve-ski
         '$jsEntrypointExtension';
     args = dart2JsArgs.toList()
       ..addAll([
+        '--packages=$multiRootScheme:///.dart_tool/package_config.json',
+        '--multi-root-scheme=$multiRootScheme',
+        '--multi-root=${scratchSpace.tempDir.uri.toFilePath()}',
         for (var experiment in enabledExperiments)
           '--enable-experiment=$experiment',
-        '--packages=${p.join('.dart_tool', 'package_config.json')}',
         '-o$jsOutputPath',
-        dartPath,
+        '$multiRootScheme:///$dartPath',
       ]);
   }
 
-  var dart2js = await buildStep.fetchResource(dart2JsWorkerResource);
-  var result = await dart2js.compile(args);
+  var librariesSpec = p.joinAll([sdkDir, 'lib', 'libraries.json']);
+  var result = await Process.run(
+      p.join(sdkDir, 'bin', 'dart'),
+      [
+        p.join(sdkDir, 'bin', 'snapshots', 'dart2js.dart.snapshot'),
+        '--libraries-spec=$librariesSpec',
+        ...args,
+      ],
+      workingDirectory: scratchSpace.tempDir.path);
   var jsOutputId = dartEntrypointId.changeExtension(jsEntrypointExtension);
   var jsOutputFile = scratchSpace.fileFor(jsOutputId);
-  if (result.succeeded && await jsOutputFile.exists()) {
-    log.info(result.output);
+  if (result.exitCode == 0 && await jsOutputFile.exists()) {
+    log.info('${result.stdout}\n${result.stderr}');
     var rootDir = p.dirname(jsOutputFile.path);
     var dartFile = p.basename(dartEntrypointId.path);
     var fileGlob = Glob('$dartFile.js*');
@@ -112,7 +128,8 @@ https://github.com/dart-lang/build/blob/master/docs/faq.md#how-can-i-resolve-ski
         dartEntrypointId.changeExtension(jsEntrypointSourceMapExtension);
     await _copyIfExists(jsSourceMapId, scratchSpace, buildStep);
   } else {
-    log.severe(result.output);
+    log.severe(
+        'ExitCode:${result.exitCode}\nStdOut:\n${result.stdout}\nStdErr:\n${result.stderr}');
   }
 }
 
@@ -123,3 +140,5 @@ Future<void> _copyIfExists(
     await scratchSpace.copyOutput(id, writer);
   }
 }
+
+final _resourcePool = Pool(maxWorkersPerTask);